+++ cf/test/test_Data.py
@@ -74,7 +74,7 @@ class DataTest(unittest.TestCase):
 
         self.test_only = []
 #        self.test_only = ['NOTHING!!!!!']
-        self.test_only = ['test_Data_halo']
+#        self.test_only = ['test_Data_halo']
 
 #        self.test_only = [
 #                          'test_Data_trigonometric_hyperbolic']